requirements.txt
PyJWT==2.8.0
fastapi==0.111.0
pydantic==2.7.4
SQLAlchemy==2.0.31
aiohttp==3.9.5
requests==2.32.3
python-dotenv==1.0.1
alembic==1.13.1
pytest==7.2.2
redis==4.5.4
celery==5.2.7
flower==1.2.0
asyncpg==0.29.0
psycopg2-binary==2.9.9